+++ supertokens_python/recipe/passwordless/asyncio/__init__.py
@@ -75,6 +75,18 @@ async def update_user(user_id: str,
     return await PasswordlessRecipe.get_instance().recipe_implementation.update_user(user_id=user_id, email=email, phone_number=phone_number, user_context=user_context)
 
 
+async def delete_email_for_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> UpdateUserResult:
+    if user_context is None:
+        user_context = {}
+    return await PasswordlessRecipe.get_instance().recipe_implementation.delete_email_for_user(user_id=user_id, user_context=user_context)
+
+
+async def delete_phone_number_for_user(user_id: str, user_context: Union[None, Dict[str, Any]] = None) -> UpdateUserResult:
+    if user_context is None:
+        user_context = {}
+    return await PasswordlessRecipe.get_instance().recipe_implementation.delete_phone_number_for_user(user_id=user_id, user_context=user_context)
+
+
 async def revoke_all_codes(email: Union[str, None] = None,
                            phone_number: Union[str, None] = None,
                            user_context: Union[None, Dict[str, Any]] = None) -> RevokeAllCodesResult: